+++ scriptabit/errors.py
@@ -60,3 +60,12 @@ class InvalidHabiticaDataError(Exception):
 
     def __str__(self):
         return repr(self.value)
+
+class NotFoundError(Exception):
+    """The specified Habitica item was not found error"""
+
+    def __init__(self, value):
+        self.value = value
+
+    def __str__(self):
+        return repr(self.value)